requirements.txt
# 核心依赖
requests>=2.31.0
python-dotenv>=1.0.0

# 网页自动化
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
imagehash>=4.3.1
scikit-image>=0.21.0

# AI和机器学习
google-generativeai>=0.3.0
openai>=1.3.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 报告生成
jinja2>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 测试框架
pytest>=7.4.0
pytest-html>=4.1.0

# 日志和工具
loguru>=0.7.0
click>=8.1.0
rich>=13.0.0

# Web框架
Flask>=2.3.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# 其他工具
beautifulsoup4>=4.12.0
lxml>=4.9.0 